Move files to Onboarding folder
