+++ app/components/UI/OnboardingWizard/Step7/index.tsx
@@ -14,7 +14,7 @@ import {
 } from '../../../../core/Analytics';
 import { useTheme } from '../../../../util/theme';
 
-import { OnboardingWizardModalSelectorsIDs } from '../../../../../e2e/selectors/Modals/OnboardingWizardModal.selectors';
+import { OnboardingWizardModalSelectorsIDs } from '../../../../../e2e/selectors/Onboarding/OnboardingWizardModal.selectors';
 import { useMetrics } from '../../../hooks/useMetrics';
 
 const styles = StyleSheet.create({
