+++ app/components/UI/SkipAccountSecurityModal/index.js
@@ -15,7 +15,7 @@ import PropTypes from 'prop-types';
 import { fontStyles } from '../../../styles/common';
 import { useTheme } from '../../../util/theme';
 import generateTestId from '../../../../wdio/utils/generateTestId';
-import { SkipAccountSecurityModalSelectorsIDs } from '../../../../e2e/selectors/Modals/SkipAccountSecurityModal.selectors';
+import { SkipAccountSecurityModalSelectorsIDs } from '../../../../e2e/selectors/Onboarding/SkipAccountSecurityModal.selectors';
 
 const createStyles = (colors) =>
   StyleSheet.create({
