+++ e2e/selectors/Onboarding/EnableAutomaticSecurityChecks.selectors.js
