+++ e2e/selectors/Onboarding/EnableDeviceNotificationsAlert.selectors.js
@@ -1,0 +1,4 @@
+export const EnableDeviceNotificationsAlertSelectorText = {
+  CONTAINER: 'Would Like to Send You Notifications',
+  YES_BUTTON: 'Allow',
+};
