+++ e2e/selectors/Onboarding/ExperienceEnhancerModal.selectors.js
