+++ e2e/selectors/Onboarding/OnboardingWizardModal.selectors.js
