+++ e2e/selectors/Onboarding/ProtectWalletModal.selectors.js
