+++ e2e/selectors/Onboarding/SkipAccountSecurityModal.selectors.js
