+++ e2e/selectors/Onboarding/WhatsNewModal.selectors.js
