+++ e2e/specs/onboarding/term-of-use.spec.js
@@ -1,4 +1,4 @@
-import TermsOfUseModal from '../../pages/modals/TermsOfUseModal';
+import TermsOfUseModal from '../../pages/Onboarding/TermsOfUseModal';
 import TestHelpers from '../../helpers';
 import OnboardingCarouselView from '../../pages/Onboarding/OnboardingCarouselView';
 import OnboardingView from '../../pages/Onboarding/OnboardingView';
