+++ e2e/specs/permission-systems/permission-system-delete-wallet.spec.js
@@ -8,7 +8,7 @@ import WalletView from '../../pages/wallet/WalletView';
 import Browser from '../../pages/Browser/BrowserView';
 import SettingsView from '../../pages/Settings/SettingsView';
 import TabBarComponent from '../../pages/TabBarComponent';
-import SkipAccountSecurityModal from '../../pages/modals/SkipAccountSecurityModal';
+import SkipAccountSecurityModal from '../../pages/Onboarding/SkipAccountSecurityModal';
 import ConnectedAccountsModal from '../../pages/Browser/ConnectedAccountsModal';
 import DeleteWalletModal from '../../pages/Settings/SecurityAndPrivacy/DeleteWalletModal';
 import LoginView from '../../pages/LoginView';
@@ -17,7 +17,7 @@ import { loginToApp } from '../../viewHelper';
 import FixtureBuilder from '../../fixtures/fixture-builder';
 import { withFixtures } from '../../fixtures/fixture-helper';
 import MetaMetricsOptIn from '../../pages/Onboarding/MetaMetricsOptInView';
-import ProtectYourWalletModal from '../../pages/modals/ProtectYourWalletModal';
+import ProtectYourWalletModal from '../../pages/Onboarding/ProtectYourWalletModal';
 import OnboardingSuccessView from '../../pages/Onboarding/OnboardingSuccessView';
 import Assertions from '../../utils/Assertions';
 import ToastModal from '../../pages/modals/ToastModal';
