+++ e2e/utils/Matchers.js
@@ -159,7 +159,7 @@ class Matchers {
    * Get system dialogs in the system-level (e.g. permissions, alerts, etc.), by text.
    *
    * @param {string} text - Match elements with the specified text
-   * @return {Promise<Detox.SystemElement>} - Resolves to the located element
+   * @return {Promise<Detox.IndexableSystemElement>} - Resolves to the located element
    */
   static async getSystemElementByText(text) {
     return system.element(by.system.label(text));
