+++ e2e/viewHelper.js
@@ -1,7 +1,7 @@
 'use strict';
 
-import EnableAutomaticSecurityChecksView from './pages/modals/EnableAutomaticSecurityChecksView';
-import EnableDeviceNotificationsAlert from './pages/EnableDeviceNotificationsAlert';
+import EnableAutomaticSecurityChecksView from './pages/Onboarding/EnableAutomaticSecurityChecksView';
+import EnableDeviceNotificationsAlert from './pages/Onboarding/EnableDeviceNotificationsAlert';
 import ImportWalletView from './pages/Onboarding/ImportWalletView';
 import MetaMetricsOptIn from './pages/Onboarding/MetaMetricsOptInView';
 import NetworkEducationModal from './pages/Network/NetworkEducationModal';
@@ -9,19 +9,19 @@ import NetworkListModal from './pages/Network/NetworkListModal';
 import NetworkView from './pages/Settings/NetworksView';
 import OnboardingView from './pages/Onboarding/OnboardingView';
 import OnboardingCarouselView from './pages/Onboarding/OnboardingCarouselView';
-import OnboardingWizardModal from './pages/modals/OnboardingWizardModal';
+import OnboardingWizardModal from './pages/Onboarding/OnboardingWizardModal';
 import SettingsView from './pages/Settings/SettingsView';
 import WalletView from './pages/wallet/WalletView';
 import Accounts from '../wdio/helpers/Accounts';
-import SkipAccountSecurityModal from './pages/modals/SkipAccountSecurityModal';
-import ProtectYourWalletModal from './pages/modals/ProtectYourWalletModal';
+import SkipAccountSecurityModal from './pages/Onboarding/SkipAccountSecurityModal';
+import ProtectYourWalletModal from './pages/Onboarding/ProtectYourWalletModal';
 import CreatePasswordView from './pages/Onboarding/CreatePasswordView';
 import ProtectYourWalletView from './pages/Onboarding/ProtectYourWalletView';
 import OnboardingSuccessView from './pages/Onboarding/OnboardingSuccessView';
 
 import TestHelpers from './helpers';
 
-import TermsOfUseModal from './pages/modals/TermsOfUseModal';
+import TermsOfUseModal from './pages/Onboarding/TermsOfUseModal';
 import TabBarComponent from './pages/TabBarComponent';
 import LoginView from './pages/LoginView';
 import { getGanachePort } from './fixtures/utils';
@@ -76,7 +76,7 @@ export const skipNotificationsDeviceSettings = async () => {
     await Assertions.checkIfVisible(
       EnableDeviceNotificationsAlert.stepOneContainer,
     );
-    await EnableDeviceNotificationsAlert.tapOnNotEnableDeviceNotificationsButton();
+    await EnableDeviceNotificationsAlert.tapOnEnableDeviceNotificationsButton();
     await Assertions.checkIfNotVisible(
       EnableDeviceNotificationsAlert.stepOneContainer,
     );
