+++ wdio/screen-objects/EnableSecurityChecksScreen.js
@@ -1,6 +1,6 @@
 import Gestures from '../helpers/Gestures';
 import Selectors from '../helpers/Selectors';
-import { EnableAutomaticSecurityChecksIDs } from '../../e2e/selectors/Modals/EnableAutomaticSecurityChecks.selectors';
+import { EnableAutomaticSecurityChecksIDs } from '../../e2e/selectors/Onboarding/EnableAutomaticSecurityChecks.selectors';
 
 class EnableAutomaticSecurityChecksScreen {
   get noThanksButton() {
