--- wdio/screen-objects/testIDs/Components/SkipAccountSecurityModalTestIds.js
@@ -1,5 +1,0 @@
-export const ANDROID_SKIP_BACKUP_BUTTON_ID = 'skip-backup-text';
-export const iOS_SKIP_BACKUP_BUTTON_ID = 'skip-backup-check';
-export const SKIP_BACKUP_MODAL = 'skip-backup-modal';
-export const SKIP_BUTTON = 'Skip-button';
-export const CANCEL_BUTTON = 'Cancel-Skip';
