--- wdio/screen-objects/testIDs/Components/TermsOfUse.testIds.js
@@ -1,10 +1,0 @@
-export const TERMS_OF_USE_SCREEN_ID = 'terms-of-use-screen-id';
-
-export const TERMS_OF_USE_ACCEPT_BUTTON_ID = 'terms-of-use-accept-button-id';
-
-export const TERMS_OF_USE_SCROLL_END_ARROW_BUTTON_ID =
-  'terms-of-use-scroll-end-arrow-button-id';
-
-export const TERMS_OF_USE_WEBVIEW_ID = 'terms-of-use-webview-id';
-
-export const TERMS_OF_USE_CHECKBOX_ICON_ID = 'terms-of-use-checkbox';
