--- wdio/screen-objects/testIDs/Screens/EnableDeviceNotificationsChecksAlert.testIds.js
@@ -1,6 +1,0 @@
-export const ENABLE_DEVICE_NOTIFICATION_NO_THANKS_BUTTON_ID =
-  "Don't Allow";
-export const ENABLE_DEVICE_NOTIFICATIONS_CONTAINER_ID =
-  'Would Like to Send You Notifications';
-export const ENABLE_DEVICE_NOTIFICATIONS_YES_BUTTON_ID = 'Allow';
-
